ggml-zdnn: move mulmat forward around
