+++ ggml/src/ggml-zdnn/mmf.cpp
@@ -1,7 +1,11 @@
 #include "ggml.h"
 #include "mmf.hpp"
 
-void ggml_zdnn_mul_mat_f(ggml_backend_zdnn_context * ctx, const ggml_tensor * src0, const ggml_tensor * src1, ggml_tensor * dst) {
+void ggml_zdnn_mul_mat_f(
+    const ggml_backend_zdnn_context * ctx,
+    const               ggml_tensor * src0,
+    const               ggml_tensor * src1,
+                        ggml_tensor * dst) {
     GGML_TENSOR_BINARY_OP_LOCALS;
 
     const enum ggml_type type = src0->type;
